requirements.txt
paho-mqtt==1.6.1
wakeonlan==3.1.0
PyYAML==6.0.1